dom/nodes/insertion-effects/Node-appendChild-script-and-style-from-fragment.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting a script and a style from a fragment</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<body>
<script>
const happened = []
const style = document.createElement("style");
let styleSheet = null;
test(() => {
  style.appendChild(new Text("body {}"));
  const script = document.createElement("script");
  script.textContent = `
    happened.push(style.sheet);
    style.appendChild(new Text("body {}"));
    happened.push(style.sheet.cssRules.length);
    styleSheet = style.sheet;
`;

  const df = document.createDocumentFragment();
  df.appendChild(script);
  df.appendChild(style);

  assert_array_equals(happened, []);
  document.body.appendChild(df);
  assert_array_equals(happened, [null, 2]);
  assert_not_equals(style.sheet, styleSheet, "style sheet was created only once");
});
</script>
